+++ include/lbann/data_readers/data_reader_hdf5_legacy.hpp
@@ -141,18 +141,19 @@ protected:
                            hsize_t filespace,
                            int rank,
                            TensorDataType* sample);
-  void
-  read_hdf5_sample(int data_id, TensorDataType* sample, TensorDataType* labels);
+  void read_hdf5_sample(uint64_t data_id,
+                        TensorDataType* sample,
+                        TensorDataType* labels);
   // void set_defaults() override;
-  void load_sample(conduit::Node& node, int data_id);
-  bool fetch_datum(CPUMat& X, int data_id, int mb_idx) override;
-  void fetch_datum_conduit(Mat& X, int data_id);
+  void load_sample(conduit::Node& node, uint64_t data_id);
+  bool fetch_datum(CPUMat& X, uint64_t data_id, uint64_t mb_idx) override;
+  void fetch_datum_conduit(Mat& X, uint64_t data_id);
   bool fetch_data_field(data_field_type data_field,
                         CPUMat& Y,
-                        int data_id,
-                        int mb_idx) override;
-  bool fetch_label(CPUMat& Y, int data_id, int mb_idx) override;
-  bool fetch_response(CPUMat& Y, int data_id, int mb_idx) override;
+                        uint64_t data_id,
+                        uint64_t mb_idx) override;
+  bool fetch_label(CPUMat& Y, uint64_t data_id, uint64_t mb_idx) override;
+  bool fetch_response(CPUMat& Y, uint64_t data_id, uint64_t mb_idx) override;
   hid_t get_hdf5_data_type() const;
   conduit::DataType get_conduit_data_type(conduit::index_t num_elements) const;
