+++ include/lbann/data_readers/data_reader_numpy_npz_conduit.hpp
@@ -75,14 +75,17 @@ public:
   {
     return m_num_response_features;
   }
-  const std::vector<El::Int> get_data_dims() const override { return m_data_dims; }
+  const std::vector<El::Int> get_data_dims() const override
+  {
+    return m_data_dims;
+  }
 
 protected:
   void do_preload_data_store() override;
 
-  bool fetch_datum(CPUMat& X, int data_id, int mb_idx) override;
-  bool fetch_label(CPUMat& Y, int data_id, int mb_idx) override;
-  bool fetch_response(CPUMat& Y, int data_id, int mb_idx) override;
+  bool fetch_datum(CPUMat& X, uint64_t data_id, uint64_t mb_idx) override;
+  bool fetch_label(CPUMat& Y, uint64_t data_id, uint64_t mb_idx) override;
+  bool fetch_response(CPUMat& Y, uint64_t data_id, uint64_t mb_idx) override;
 
   /// Number of samples.
   int m_num_samples = 0;
@@ -111,13 +114,14 @@ protected:
                                 std::unordered_set<int>& label_classes);
 
   void load_conduit_node(const std::string filename,
-                         int data_id,
+                         uint64_t data_id,
                          conduit::Node& output,
                          bool reset = true);
 
   std::unordered_map<int, std::map<std::string, cnpy::NpyArray>> m_npz_cache;
 
-  void load_npz(const std::string filename, int data_id, conduit::Node& node);
+  void
+  load_npz(const std::string filename, uint64_t data_id, conduit::Node& node);
 };
 
 } // namespace lbann
