+++ include/lbann/data_readers/data_reader_sample_list_impl.hpp
@@ -77,7 +77,7 @@ void data_reader_sample_list<SampleListT>::shuffle_indices(rng_gen& gen)
 {
   generic_data_reader::shuffle_indices(gen);
   // BVE FIXME get_mini_batch_size()
-  auto mini_batch_size = (size_t) 1;
+  auto mini_batch_size = (uint64_t)1;
   if (trainer_exists()) {
     mini_batch_size = get_trainer().get_max_mini_batch_size();
   }
